backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pillow==10.3.0
opencv-python==4.9.0.80
numpy==1.26.4
ultralytics==8.2.0
reportlab==4.1.0
aiofiles==23.2.1
pyttsx3==2.90
requests==2.31.0
scipy==1.13.0
matplotlib==3.8.4
scikit-learn>=1.4.0
google-generativeai>=0.5.0
pymysql>=1.1.0
python-dotenv>=1.0.0
torch
torchvision
